+++ integ-tests/SutProject.Tests/Customization/CustomServiceProviderAttribute.cs
@@ -1,12 +1,12 @@
-using DependencyModules.xUnit.Attributes.Interfaces;
+using DependencyModules.Testing.Attributes.Interfaces;
 using Microsoft.Extensions.DependencyInjection;
-using Xunit.v3;
 
 namespace SutProject.Tests.Customization;
 
 public class CustomServiceProviderAttribute : Attribute, IServiceProviderBuilderAttribute {
-    
-    public IServiceProvider BuildServiceProvider(IXunitTestMethod testCaseContext, IServiceCollection serviceCollection) {
+
+    public IServiceProvider BuildServiceProvider(
+        ITestMethodContext testMethod, IServiceCollection serviceCollection) {
         serviceCollection.AddSingleton<ICustomTestDependency, CustomTestDependency>();
         return serviceCollection.BuildServiceProvider();
     }